pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "Leandro Freitas", email = "vpc2112@gmail.com"},
]
dependencies = [
    "pandas>=1.4.3",
    "influxdb-client>=1.31.0",
    "datetime>=4.5",
    "influxdb>=5.3.1",
    "mysql-connector-python>=8.0.30",
    "reactivex>=4.0.4",
]
requires-python = ">=3.10"
license = {text = "MIT"}

[build-system]
requires = ["pdm-pep517>=1.0.0"]
build-backend = "pdm.pep517.api"


[tool]
[tool.pdm]
